nebulon-backend/requirements.txt
fastapi
uvicorn
solana
anchorpy
httpx
pydantic
python-dotenv
